+++ src/rlang.c
@@ -1,0 +1,17 @@
+#include "rlang/rlang.h"
+
+#include "rlang/utils.c"
+#include "rlang/sym.c"
+#include "rlang/formula.c"
+#include "rlang/quo.c"
+#include "rlang/env.c"
+#include "rlang/node.c"
+#include "rlang/lang.c"
+#include "rlang/sexp.c"
+#include "rlang/vector.c"
+#include "rlang/vector-chr.c"
+#include "rlang/vector-lgl.c"
+#include "rlang/vector-list.c"
+#include "rlang/eval.c"
+#include "rlang/cnd.c"
+#include "rlang/stack.c"
