+++ src/rlang/stack.c
