+++ src/rlang/stack.h
